perfect some details

+++ src/zhou/servlet/InOrderInfoServlet.java
@@ -34,9 +34,15 @@ public class InOrderInfoServlet extends HttpServlet {
 		request.setCharacterEncoding("UTF-8");
 		response.setCharacterEncoding("UTF-8");
 		//PrintWriter out = response.getWriter();
+		
+		if (request.getSession().getAttribute("admin") == null) {
+			response.sendRedirect("pages/login/login.html");
+			return;
+		}
+		
 		System.out.println("------------------------------------------------");
 		System.out.println("In Here: InOrderInfoServlet");
-
+		
 		System.out.println(request.getQueryString());
 		
 		DeleteOrderInInfo(request);
